+++ src/Oro/Bundle/ApiBundle/Tests/Unit/Processor/Shared/JsonApi/SetOperationFlagsTest.php
@@ -26,6 +26,7 @@ class SetOperationFlagsTest extends FormProcessorTestCase
 
         self::assertFalse($this->context->has(SetOperationFlags::UPDATE_FLAG));
         self::assertFalse($this->context->has(SetOperationFlags::UPSERT_FLAG));
+        self::assertFalse($this->context->has(SetOperationFlags::VALIDATE_FLAG));
     }
 
     public function testProcessWhenRequestDataDoesNotContainMeta(): void
@@ -35,6 +36,7 @@ class SetOperationFlagsTest extends FormProcessorTestCase
 
         self::assertFalse($this->context->has(SetOperationFlags::UPDATE_FLAG));
         self::assertFalse($this->context->has(SetOperationFlags::UPSERT_FLAG));
+        self::assertFalse($this->context->has(SetOperationFlags::VALIDATE_FLAG));
     }
 
     public function testProcessWhenRequestDataContainsNotArrayMeta(): void
@@ -44,6 +46,7 @@ class SetOperationFlagsTest extends FormProcessorTestCase
 
         self::assertFalse($this->context->has(SetOperationFlags::UPDATE_FLAG));
         self::assertFalse($this->context->has(SetOperationFlags::UPSERT_FLAG));
+        self::assertFalse($this->context->has(SetOperationFlags::VALIDATE_FLAG));
     }
 
     public function testProcessWhenRequestDataContainsEmptyMeta(): void
@@ -53,6 +56,7 @@ class SetOperationFlagsTest extends FormProcessorTestCase
 
         self::assertFalse($this->context->has(SetOperationFlags::UPDATE_FLAG));
         self::assertFalse($this->context->has(SetOperationFlags::UPSERT_FLAG));
+        self::assertFalse($this->context->has(SetOperationFlags::VALIDATE_FLAG));
     }
 
     public function testProcessWhenRequestDataContainsUpdateFlagInMeta(): void
@@ -62,6 +66,7 @@ class SetOperationFlagsTest extends FormProcessorTestCase
 
         self::assertTrue($this->context->get(SetOperationFlags::UPDATE_FLAG));
         self::assertFalse($this->context->has(SetOperationFlags::UPSERT_FLAG));
+        self::assertFalse($this->context->has(SetOperationFlags::VALIDATE_FLAG));
     }
 
     public function testProcessWhenRequestDataContainsUpsertFlagInMeta(): void
@@ -71,6 +76,17 @@ class SetOperationFlagsTest extends FormProcessorTestCase
 
         self::assertFalse($this->context->has(SetOperationFlags::UPDATE_FLAG));
         self::assertTrue($this->context->get(SetOperationFlags::UPSERT_FLAG));
+        self::assertFalse($this->context->has(SetOperationFlags::VALIDATE_FLAG));
+    }
+
+    public function testProcessWhenRequestDataContainsValidateFlagInMeta(): void
+    {
+        $this->context->setRequestData(['data' => ['type' => 'test', 'meta' => ['validate' => true]]]);
+        $this->processor->process($this->context);
+
+        self::assertFalse($this->context->has(SetOperationFlags::UPDATE_FLAG));
+        self::assertFalse($this->context->has(SetOperationFlags::UPSERT_FLAG));
+        self::assertTrue($this->context->get(SetOperationFlags::VALIDATE_FLAG));
     }
 
     public function testProcessWhenRequestDataContainsInvalidUpdateFlagInMeta(): void
@@ -80,6 +96,7 @@ class SetOperationFlagsTest extends FormProcessorTestCase
 
         self::assertFalse($this->context->has(SetOperationFlags::UPDATE_FLAG));
         self::assertFalse($this->context->has(SetOperationFlags::UPSERT_FLAG));
+        self::assertFalse($this->context->has(SetOperationFlags::VALIDATE_FLAG));
         self::assertEquals(
             [
                 Error::createValidationError(
@@ -98,12 +115,32 @@ class SetOperationFlagsTest extends FormProcessorTestCase
 
         self::assertFalse($this->context->has(SetOperationFlags::UPDATE_FLAG));
         self::assertFalse($this->context->has(SetOperationFlags::UPSERT_FLAG));
+        self::assertFalse($this->context->has(SetOperationFlags::VALIDATE_FLAG));
         self::assertEquals(
             [
                 Error::createValidationError(
                     Constraint::VALUE,
                     'This value should be a boolean or an array of strings.'
                 )->setSource(ErrorSource::createByPointer('/meta/upsert'))
+            ],
+            $this->context->getErrors()
+        );
+    }
+
+    public function testProcessWhenRequestDataContainsInvalidValidateFlagInMeta(): void
+    {
+        $this->context->setRequestData(['data' => ['type' => 'test', 'meta' => ['validate' => 'test']]]);
+        $this->processor->process($this->context);
+
+        self::assertFalse($this->context->has(SetOperationFlags::UPDATE_FLAG));
+        self::assertFalse($this->context->has(SetOperationFlags::UPSERT_FLAG));
+        self::assertFalse($this->context->has(SetOperationFlags::VALIDATE_FLAG));
+        self::assertEquals(
+            [
+                Error::createValidationError(
+                    Constraint::VALUE,
+                    'This value should be a boolean.'
+                )->setSource(ErrorSource::createByPointer('/meta/validate'))
             ],
             $this->context->getErrors()
         );